[#15] rename and mv backend files under fastapi_server
