+++ fastapi_server/initializer.py
@@ -1,6 +1,7 @@
 import os
 import inject
-from database.database import Database, PortfolioRepositoryService
+from fastapi_server.database.database import Database
+from fastapi_server.repository.portfolio_repository_service import PortfolioRepositoryService
 from prophecy import ProphetModel
 from trader import Trader
