+++ fastapi_server/repository/__init__.py
@@ -1,0 +1,5 @@
+"""
+Author: zed.ai
+Reviewer:
+2024.05.22
+"""
